➖ remove useless dependency
