+++ jest.config.ts
@@ -1,0 +1,15 @@
+import type { Config } from "jest";
+import nextJest from "next/jest.js";
+
+const createJestConfig = nextJest({
+  dir: "./",
+});
+
+const config: Config = {
+  // Add more setup options before each test is run
+  setupFilesAfterEnv: ["<rootDir>/jest.setup.ts"],
+  preset: "ts-jest",
+  testEnvironment: "jest-environment-jsdom",
+};
+
+export default createJestConfig(config);
